Created authenticate folder in webapp and moved related files
