+++ ai4good/webapp/authenticate/webapp_file.py
@@ -10,12 +10,10 @@ from flask_login import logout_user
 from werkzeug.urls import url_parse
 
 from ai4good.webapp.apps import db_sqlalchemy
-from ai4good.forms import LoginForm
-from ai4good.forms import RegistrationForm
-from ai4good.models_file import User
+from ai4good.webapp.authenticate.forms import LoginForm, RegistrationForm
+from ai4good.webapp.authenticate.models_file import User
 
 server_bp = Blueprint('main', __name__)
-
 
 @server_bp.route('/')
 def index():
